Decoding allows structs to end with a flat positional field
Some kicad structures are a set of fixed positional values followed by
a variable number of positional values within a single tuple. To
accommodate this we allow a struct to end with a positional field that
has the "flat" field set, thus consuming all of the remaining values
into its slice.

This allows us to decode "layer" tuples, which have optional flag tokens
at the end ("hide", for example).